+++ projects/packages/videopress/src/client/state/constants.js
@@ -12,6 +12,7 @@ export const WP_REST_API_MEDIA_ENDPOINT = 'wp/v2/media';
 export const WP_REST_API_VIDEOPRESS_ENDPOINT = 'wpcom/v2/videopress';
 export const WP_REST_API_VIDEOPRESS_META_ENDPOINT = 'wpcom/v2/videopress/meta';
 export const WP_REST_API_VIDEOPRESS_PLAYBACK_TOKEN_ENDPOINT = 'wpcom/v2/videopress/playback-jwt';
+export const WP_REST_API_VIDEOPRESS_SETTINGS_ENDPOINT = 'videopress/v1/settings';
 export const REST_API_SITE_PURCHASES_ENDPOINT = 'my-jetpack/v1/site/purchases';
 export const REST_API_SITE_INFO_ENDPOINT = 'videopress/v1/site';
 
@@ -58,6 +59,8 @@ export const SET_USERS_PAGINATION = 'SET_USERS_PAGINATION';
 export const SET_UPDATING_VIDEO_POSTER = 'SET_UPDATING_VIDEO_POSTER';
 export const UPDATE_VIDEO_POSTER = 'UPDATE_VIDEO_POSTER';
 
+export const SET_VIDEOPRESS_SETTINGS = 'SET_VIDEOPRESS_SETTINGS';
+
 /*
  * Accepted file extensions
  */
@@ -80,3 +83,7 @@ export const VIDEO_PRIVACY_LEVELS = [
 	VIDEO_PRIVACY_LEVEL_PRIVATE,
 	VIDEO_PRIVACY_LEVEL_SITE_DEFAULT,
 ];
+
+export const VIDEO_RATING_G = 'G';
+export const VIDEO_RATING_PG_13 = 'PG-13';
+export const VIDEO_RATING_R_17 = 'R-17';