+++ projects/plugins/jetpack/class.jetpack-gutenberg.php
@@ -712,8 +712,9 @@ class Jetpack_Gutenberg {
 		if ( Jetpack::is_module_active( 'publicize' ) && function_exists( 'publicize_init' ) ) {
 			$publicize               = publicize_init();
 			$initial_state['social'] = array(
-				'sharesData'  => $publicize->get_publicize_shares_info( $blog_id ),
-				'hasPaidPlan' => $publicize->has_paid_plan(),
+				'sharesData'                  => $publicize->get_publicize_shares_info( $blog_id ),
+				'hasPaidPlan'                 => $publicize->has_paid_plan(),
+				'isEnhancedPublishingEnabled' => $publicize->is_enhanced_publishing_enabled( $blog_id ),
 			);
 		}