+++ projects/plugins/jetpack/modules/contact-form/js/grunion-admin.js
@@ -29,41 +29,6 @@ jQuery( function ( $ ) {
 			.append( jetpack_empty_spam_feedbacks_button_container );
 	}
 
-	$( document ).on( 'click', '#jetpack-check-feedback-spam:not(.button-disabled)', function ( e ) {
-		e.preventDefault();
-
-		$( '#jetpack-check-feedback-spam:not(.button-disabled)' ).addClass( 'button-disabled' );
-		$( '.jetpack-check-feedback-spam-spinner' ).addClass( 'spinner' ).show();
-		grunion_check_for_spam( 0, 100 );
-	} );
-
-	function grunion_check_for_spam( offset, limit ) {
-		var nonceName = $( '#jetpack-check-feedback-spam' ).data( 'nonce-name' );
-		var nonce = $( '#' + nonceName ).attr( 'value' );
-		var failureUrl = $( '#jetpack-check-feedback-spam' ).data( 'failure-url' );
-
-		var requestOptions = {
-			action: 'grunion_recheck_queue',
-			offset: offset,
-			limit: limit,
-		};
-		requestOptions[ nonceName ] = nonce;
-
-		$.post( ajaxurl, requestOptions )
-			.fail( function ( result ) {
-				// An error is only returned in the case of a missing nonce or invalid permissions, so we don't need the actual error message.
-				window.location.href = failureUrl;
-				return;
-			} )
-			.done( function ( result ) {
-				if ( result.processed < limit ) {
-					window.location.reload();
-				} else {
-					grunion_check_for_spam( offset + limit, limit );
-				}
-			} );
-	}
-
 	var initial_spam_count = 0;
 	var deleted_spam_count = 0;
 
@@ -121,6 +86,63 @@ jQuery( function ( $ ) {
 				}
 			} );
 	}
+
+	// Async handling for response table actions
+	$( document ).ready( function () {
+		function updateStatus( postId, status, indicatorColor ) {
+			$.post(
+				ajaxurl,
+				{
+					action: 'grunion_ajax_spam',
+					post_id: postId,
+					make_it: status,
+					sub_menu: jQuery( '.subsubsub .current' ).attr( 'href' ),
+					_ajax_nonce: window.__grunionPostStatusNonce,
+				},
+				function ( response ) {
+					$( '#post-' + postId )
+						.css( { backgroundColor: indicatorColor } )
+						.fadeOut( 350, function () {
+							$( this ).remove();
+							$( '.subsubsub' ).html( response );
+						} );
+				}
+			);
+		}
+
+		$( 'tr.type-feedback .row-actions a' ).click( function ( e ) {
+			e.preventDefault();
+
+			var postRowId = $( e.target ).closest( 'tr.type-feedback' ).attr( 'id' );
+			var match = postRowId.match( /^post\-(\d+)/ );
+
+			if ( ! match ) {
+				return;
+			}
+
+			var postId = parseInt( match[ 1 ], 10 );
+
+			if ( $( e.target ).parent().hasClass( 'spam' ) ) {
+				e.preventDefault();
+				updateStatus( postId, 'spam', '#FF7979' );
+			}
+
+			if ( $( e.target ).parent().hasClass( 'trash' ) ) {
+				e.preventDefault();
+				updateStatus( postId, 'trash', '#FF7979' );
+			}
+
+			if ( $( e.target ).parent().hasClass( 'unspam' ) ) {
+				e.preventDefault();
+				updateStatus( postId, 'ham', '#59C859' );
+			}
+
+			if ( $( e.target ).parent().hasClass( 'untrash' ) ) {
+				e.preventDefault();
+				updateStatus( postId, 'publish', '#59C859' );
+			}
+		} );
+	} );
 
 	// Handle export
 	$( document ).on( 'click', '#jetpack-export-feedback', function ( e ) {